:white_check_mark: Add unit test for inmemory cache (#270)

+++ pkg/configs/configuration.go
@@ -114,7 +114,7 @@ type ConfigurationCacheRedis struct {
 
 // ConfigurationCacheInmemory ...
 type ConfigurationCacheInmemory struct {
-	Size int64 `yaml:"size"`
+	Size int `yaml:"size"`
 }
 
 // ConfigurationCacheDatabase ...
